seperate test for bytecode
