+++ zkevm-circuits/src/bytecode_circuit.rs
@@ -3,3 +3,7 @@
 /// Bytecode unroller
 pub mod bytecode_unroller;
 pub(crate) mod param;
+/// bytecode circuit test
+#[cfg(any(feature = "test", test))]
+mod test;
+pub use test::BytecodeCircuit;
